test: [#1251] unify test runner for sqlite and mysql drivers
In the tracker-core package.
